+++ src/app/[variants]/(main)/agent/integration/const.ts
@@ -23,10 +23,10 @@ export const INTEGRATION_PROVIDERS: IntegrationProvider[] = [
     description: 'Connect this assistant to Discord server for channel chat and direct messages.',
     docsLink: 'https://discord.com/developers/docs/intro',
     fieldTags: {
-      appId: 'APPLICATION ID',
-      publicKey: 'PUBLIC KEY',
-      token: 'TOKEN',
-      webhook: 'INTERACTIONS ENDPOINT URL',
+      appId: 'Application ID',
+      publicKey: 'Public Key',
+      token: 'Bot Token',
+      webhook: 'Interactions Endpoint URL',
     },
     icon: SiDiscord,
     id: 'discord',
